v 0.2 Major update! It works!
Significant improvements to the GUI such that many of the images show up. Fixed more coding related to the find_samples function such that it works now. Added functions for "go to xyzr position" and "take snapshot with current GUI settings". Significant restructuring to move the microscope control to the GUI thread, such that the connection will be available for all buttons rather than connecting per button.