Major update - data collection
Data collection is now available by way of two options. First, use a standard workflow with a bounding box to generate a single collection - this is intended for a single Zstack. Second, there is a multi-angle collection option that interpolates the location of the sample from a set of bounding boxes. This can serve as a template for other types of functions for more complex collections. Each collection is saved to a separate subfolder with major or minor modifications to the workflow file sent.